extend default endpoints to bcf api
