+++ lib/api/utilities/endpoints/create.rb
@@ -27,12 +27,22 @@
 #++
 
 module API
-  module V3
-    module Utilities
-      module Endpoints
-        class Update < API::Utilities::Endpoints::Update
-          include V3Deductions
-          include V3PresentSingle
+  module Utilities
+    module Endpoints
+      class Create < Modify
+        def default_instance_generator(_model)
+          ->(_params) do
+          end
+        end
+
+        def success_status
+          :created
+        end
+
+        private
+
+        def update_or_create
+          "Create"
         end
       end
     end
