+++ lib/api/utilities/endpoints/update.rb
@@ -27,12 +27,19 @@
 #++
 
 module API
-  module V3
-    module Utilities
-      module Endpoints
-        class Update < API::Utilities::Endpoints::Update
-          include V3Deductions
-          include V3PresentSingle
+  module Utilities
+    module Endpoints
+      class Update < Modify
+        def default_instance_generator(model)
+          ->(_params) do
+            instance_variable_get("@#{model.name.demodulize.underscore}")
+          end
+        end
+
+        private
+
+        def update_or_create
+          "Update"
         end
       end
     end
