+++ lib/api/v3/utilities/endpoints/index.rb
@@ -30,122 +30,7 @@ module API
   module V3
     module Utilities
       module Endpoints
-        class Index
-          include ::API::Utilities::PageSizeHelper
-
-          def initialize(model:,
-                         api_name: model.name.demodulize,
-                         scope: nil,
-                         render_representer: nil)
-            self.model = model_class(model)
-            self.scope = scope
-            self.api_name = api_name
-            self.render_representer = render_representer || deduce_render_representer
-          end
-
-          def mount
-            index = self
-
-            -> do
-              query = index.parse(params)
-
-              self_path = api_v3_paths.send(index.self_path)
-              base_scope = index.scope ? instance_exec(&index.scope) : index.model
-
-              index.render(query, params, self_path, base_scope)
-            end
-          end
-
-          def parse(params)
-            ParamsToQueryService
-              .new(model, User.current)
-              .call(params)
-          end
-
-          def render(query, params, self_path, base_scope)
-            if query.valid?
-              render_success(query, params, self_path, base_scope)
-            else
-              render_error(query)
-            end
-          end
-
-          def self_path
-            api_name.underscore.pluralize
-          end
-
-          attr_accessor :model,
-                        :api_name,
-                        :scope,
-                        :render_representer
-
-          private
-
-          def render_success(query, params, self_path, base_scope)
-            results = merge_scopes(base_scope, query.results)
-
-            if paginated_representer?
-              render_paginated_success(results, params, self_path)
-            else
-              render_unpaginated_success(results, self_path)
-            end
-          end
-
-          def render_paginated_success(results, params, self_path)
-            render_representer
-              .new(results,
-                   self_path,
-                   page: to_i_or_nil(params[:offset]),
-                   per_page: resolve_page_size(params[:pageSize]),
-                   current_user: User.current)
-          end
-
-          def render_unpaginated_success(results, self_path)
-            render_representer
-              .new(results,
-                   self_path,
-                   current_user: User.current)
-          end
-
-          def paginated_representer?
-            render_representer.ancestors.include?(::API::Decorators::OffsetPaginatedCollection)
-          end
-
-          def render_error(query)
-            raise ::API::Errors::InvalidQuery.new(query.errors.full_messages)
-          end
-
-          def deduce_render_representer
-            "::API::V3::#{deduce_namespace}::#{api_name}CollectionRepresenter".constantize
-          end
-
-          def deduce_namespace
-            api_name.pluralize
-          end
-
-          def model_class(scope)
-            if scope.is_a? Class
-              scope
-            else
-              scope.model
-            end
-          end
-
-          def with_based_scope(other_scope)
-            if scope
-              merge_scopes(scope, other_scope)
-            else
-              other_scope
-            end
-          end
-
-          def merge_scopes(scope_a, scope_b)
-            if scope_a.is_a? Class
-              scope_b
-            else
-              scope_a.merge(scope_b)
-            end
-          end
+        class Index < API::Utilities::Endpoints::Index
         end
       end
     end
