+++ lib/api/v3/utilities/endpoints/show.rb
@@ -30,57 +30,8 @@ module API
   module V3
     module Utilities
       module Endpoints
-        class Show
-          def default_instance_generator(model)
-            ->(_params) do
-              instance_variable_get("@#{model.name.demodulize.underscore}")
-            end
-          end
-
-          def initialize(model:,
-                         api_name: model.name.demodulize,
-                         render_representer: nil,
-                         instance_generator: default_instance_generator(model))
-
-            self.model = model
-            self.api_name = api_name
-            self.instance_generator = instance_generator
-            self.render_representer = render_representer || deduce_render_representer
-          end
-
-          def mount
-            show = self
-
-            -> do
-              show.render(instance_exec(params, &show.instance_generator))
-            end
-          end
-
-          def render(instance)
-            render_representer
-              .create(instance,
-                      current_user: User.current,
-                      embed_links: true)
-          end
-
-          def self_path
-            api_name.underscore.pluralize
-          end
-
-          attr_accessor :model,
-                        :api_name,
-                        :instance_generator,
-                        :render_representer
-
-          private
-
-          def deduce_render_representer
-            "::API::V3::#{deduce_namespace}::#{api_name}Representer".constantize
-          end
-
-          def deduce_namespace
-            api_name.pluralize
-          end
+        class Show < API::Utilities::Endpoints::Show
+          include V3Deductions
         end
       end
     end
